requirements.txt
numpy
torch
torchvision
dvc
paramiko
bpython
natsorted
tensorboardx
ray
psutil
jupyter
setproctitle
hyperopt
pdbpp
